requirements.txt
numpy
opencv-python
scikit-image
scikit-learn
PySide6
Pillow
scipy
SimpleITK
colour-science
colormath
OpenColorIO
imagequant
realesrgan
basicsr
facexlib
gfpgan
codeformer
lpips
torch
torchvision
tqdm
